requirements.txt
Flask~=3.1.1
MetaTrader5~=5.0.5200
python-dotenv~=1.1.1